requirements.txt
streamlit==1.45.0
numpy==1.26.4
scikit-learn==1.6.1
tensorflow==2.18.0
pillow==11.2.1
faiss-cpu == 1.8.0
